Upgrade Go. Implement user and sessions. Implement web framework. Comment out Basic Auth and update Readme. Remove trailing slash from paths.

+++ app/category/category.go
@@ -16,6 +16,7 @@ type Category struct {
 	Description string
 	Name        string
 	TextColor   string
+	UserId      int
 }
 
 func CreateCategoryTable(db *sql.DB) {
@@ -26,7 +27,12 @@ func CreateCategoryTable(db *sql.DB) {
 			description VARCHAR(150) DEFAULT '',
 			name VARCHAR(30) NOT NULL,
 			text_color CHAR(7) DEFAULT '#000000',
-			PRIMARY KEY (id)
+			user_id INT NOT NULL,
+			PRIMARY KEY (id),
+			FOREIGN KEY fk_user_id (user_id)
+				REFERENCES users (id)
+				ON DELETE CASCADE
+				ON UPDATE NO ACTION
 		);
 	`
 	log.Println("Creating the categories table.")
@@ -40,51 +46,51 @@ func CreateCategoryTable(db *sql.DB) {
 	log.Println("Categories table created.")
 }
 
-func DeleteCategory(db *sql.DB, id int) (err error) {
-	_, err = db.Exec("DELETE FROM categories WHERE id = ?;", id)
+func DeleteCategory(db *sql.DB, id int, uid int) (err error) {
+	_, err = db.Exec("DELETE FROM categories WHERE id = ? AND user_id = ?;", id, uid)
 	return
 }
 
-func GetCategory(db *sql.DB, id int) (category Category, err error) {
-	row := db.QueryRow("SELECT * FROM categories WHERE id = ?;", id)
-	err = row.Scan(&category.Id, &category.Color, &category.Description, &category.Name, &category.TextColor)
+func GetCategory(db *sql.DB, id int, uid int) (c Category, err error) {
+	row := db.QueryRow("SELECT * FROM categories WHERE id = ? AND user_id = ?;", id, uid)
+	err = row.Scan(&c.Id, &c.Color, &c.Description, &c.Name, &c.TextColor, &c.UserId)
 	return
 }
 
-func GetCategories(db *sql.DB) (categories []Category, err error) {
-	rows, err := db.Query("SELECT  * FROM categories;")
+func GetCategories(db *sql.DB, uid int) (cs []Category, err error) {
+	rows, err := db.Query("SELECT * FROM categories WHERE user_id = ?;", uid)
 	if err != nil {
 		return
 	}
 	defer rows.Close()
 	for rows.Next() {
-		category := Category{}
-		if err = rows.Scan(&category.Id, &category.Color, &category.Description, &category.Name, &category.TextColor); err != nil {
+		c := Category{}
+		if err = rows.Scan(&c.Id, &c.Color, &c.Description, &c.Name, &c.TextColor, &c.UserId); err != nil {
 			return
 		}
-		categories = append(categories, category)
+		cs = append(cs, c)
 	}
 	return
 }
 
-func InsertCategory(db *sql.DB, category Category) (err error) {
+func InsertCategory(db *sql.DB, c Category) (err error) {
 	statement, err := db.Prepare(
-		"INSERT INTO categories(color, description, name, text_color) VALUES (?, ?, ?, ?);",
+		"INSERT INTO categories(color, description, name, text_color, user_id) VALUES (?, ?, ?, ?, ?);",
 	)
 	if err != nil {
 		return
 	}
-	_, err = statement.Exec(category.Color, category.Description, category.Name, category.TextColor)
+	_, err = statement.Exec(c.Color, c.Description, c.Name, c.TextColor, c.UserId)
 	return
 }
 
-func UpdateCategory(db *sql.DB, category Category) (err error) {
+func UpdateCategory(db *sql.DB, c Category) (err error) {
 	statement, err := db.Prepare(
-		"UPDATE categories SET color = ?, description = ?, name = ?, text_color = ? WHERE id = ?;",
+		"UPDATE categories SET color = ?, description = ?, name = ?, text_color = ? WHERE id = ? AND user_id = ?;",
 	)
 	if err != nil {
 		return
 	}
-	_, err = statement.Exec(category.Color, category.Description, category.Name, category.TextColor, category.Id)
+	_, err = statement.Exec(c.Color, c.Description, c.Name, c.TextColor, c.Id, c.UserId)
 	return
 }
