+++ app/transaction/transaction.go
@@ -1,15 +1,11 @@
 package transaction
 
 import (
-	"bitbucket.org/zanvd/accountant/category"
 	"database/sql"
 	"log"
-	"time"
-)
 
-const (
-	dbDateFormat      = "2006-01-02"
-	displayDateFormat = "02. 01. 2006"
+	"bitbucket.org/zanvd/accountant/category"
+	"bitbucket.org/zanvd/accountant/convert"
 )
 
 type Transaction struct {
@@ -19,10 +15,7 @@ type Transaction struct {
 	Name            string
 	Summary         string
 	TransactionDate string
-}
-
-func CurrentDateInDisplayFormat() string {
-	return time.Now().Format(displayDateFormat)
+	UserId          int
 }
 
 func CreateTransactionsTable(db *sql.DB) {
@@ -34,10 +27,15 @@ func CreateTransactionsTable(db *sql.DB) {
 			name VARCHAR(30) NOT NULL,
 			summary VARCHAR(200) DEFAULT '',
 			transaction_date DATE NOT NULL,
+			user_id INT NOT NULL,
 			PRIMARY KEY (id),
-			FOREIGN KEY category_id_idx (category_id)
+			FOREIGN KEY fk_category_id (category_id)
 			    REFERENCES categories (id)
 			    ON DELETE RESTRICT
+				ON UPDATE NO ACTION,
+			FOREIGN KEY fk_user_id (user_id)
+				REFERENCES users (id)
+				ON DELETE CASCADE
 				ON UPDATE NO ACTION
 		);
 	`
@@ -52,92 +50,77 @@ func CreateTransactionsTable(db *sql.DB) {
 	log.Println("Transactions table created.")
 }
 
-func DbToDisplayDate(dbDate string) string {
-	return changeDateFormat(displayDateFormat, dbDateFormat, dbDate)
-}
-
-func DisplayTimeToDb(displayDate string) string {
-	return changeDateFormat(dbDateFormat, displayDateFormat, displayDate)
-}
-
-func DeleteTransaction(db *sql.DB, id int) (err error) {
-	_, err = db.Exec("DELETE FROM transactions WHERE id = ?;", id)
+func DeleteTransaction(db *sql.DB, id int, uid int) (err error) {
+	_, err = db.Exec("DELETE FROM transactions WHERE id = ? AND user_id = ?;", id, uid)
 	return
 }
 
-func GetTransaction(db *sql.DB, id int) (transaction Transaction, err error) {
+func GetTransaction(db *sql.DB, id int, uid int) (t Transaction, err error) {
 	query := `
-		SELECT t.id, t.amount, t.name, t.summary, t.transaction_date,
+		SELECT t.id, t.amount, t.name, t.summary, t.transaction_date, t.user_id,
 		       c.id, c.color, c.description, c.name, c.text_color
 		FROM transactions t
 		LEFT JOIN categories c ON c.id = t.category_id
-		WHERE t.id = ?;
+		WHERE t.id = ? AND t.user_id = ?;
 	`
-	row := db.QueryRow(query, id)
-	err = row.Scan(&transaction.Id, &transaction.Amount, &transaction.Name, &transaction.Summary,
-		&transaction.TransactionDate, &transaction.Category.Id, &transaction.Category.Color,
-		&transaction.Category.Description, &transaction.Category.Name, &transaction.Category.TextColor)
+	row := db.QueryRow(query, id, uid)
+	err = row.Scan(
+		&t.Id, &t.Amount, &t.Name, &t.Summary, &t.TransactionDate, &t.UserId,
+		&t.Category.Id, &t.Category.Color, &t.Category.Description, &t.Category.Name, &t.Category.TextColor,
+	)
 	return
 }
 
-func GetTransactions(db *sql.DB) (transactions []Transaction, err error) {
+func GetTransactions(db *sql.DB, uid int) (ts []Transaction, err error) {
 	query := `
-		SELECT t.id, t.amount, t.name, t.summary, t.transaction_date,
+		SELECT t.id, t.amount, t.name, t.summary, t.transaction_date, t.user_id,
 		       c.id, c.color, c.description, c.name, c.text_color
 		FROM transactions t
-		LEFT JOIN categories c ON c.id = t.category_id;
+		LEFT JOIN categories c ON c.id = t.category_id
+		WHERE t.user_id = ?;
 	`
-	rows, err := db.Query(query)
+	rows, err := db.Query(query, uid)
 	if err != nil {
 		return
 	}
 	defer rows.Close()
 	for rows.Next() {
-		transaction := Transaction{}
-		if err = rows.Scan(&transaction.Id, &transaction.Amount, &transaction.Name, &transaction.Summary,
-			&transaction.TransactionDate, &transaction.Category.Id, &transaction.Category.Color,
-			&transaction.Category.Description, &transaction.Category.Name, &transaction.Category.TextColor); err != nil {
+		t := Transaction{}
+		if err = rows.Scan(
+			&t.Id, &t.Amount, &t.Name, &t.Summary, &t.TransactionDate, &t.UserId,
+			&t.Category.Id, &t.Category.Color, &t.Category.Description, &t.Category.Name, &t.Category.TextColor,
+		); err != nil {
 			return
 		}
-		transactions = append(transactions, transaction)
+		ts = append(ts, t)
 	}
 	return
 }
 
-func InsertTransaction(db *sql.DB, transaction Transaction) (err error) {
-	if len(transaction.TransactionDate) == 0 {
-		transaction.TransactionDate = time.Now().UTC().Format(dbDateFormat)
+func InsertTransaction(db *sql.DB, t Transaction) (err error) {
+	if len(t.TransactionDate) == 0 {
+		t.TransactionDate = convert.CurrentDateInDbFormat()
 	}
-	query := "INSERT INTO transactions(amount, category_id, name, summary, transaction_date) VALUES (?, ?, ?, ?, ?);"
-	statement, err := db.Prepare(query)
+	statement, err := db.Prepare(`
+		INSERT INTO transactions(amount, category_id, name, summary, transaction_date, user_id)
+		VALUES (?, ?, ?, ?, ?, ?);
+	`)
 	if err != nil {
 		return
 	}
-	_, err = statement.Exec(transaction.Amount, transaction.Category.Id, transaction.Name, transaction.Summary,
-		transaction.TransactionDate)
+	_, err = statement.Exec(t.Amount, t.Category.Id, t.Name, t.Summary, t.TransactionDate, t.UserId)
 	return
 }
 
-func UpdateTransaction(db *sql.DB, transaction Transaction) (err error) {
-	query := `
+func UpdateTransaction(db *sql.DB, t Transaction) (err error) {
+	statement, err := db.Prepare(`
 		UPDATE transactions
 		SET amount = ?, category_id = ?, name = ?, summary = ?, transaction_date = ?
-		WHERE id = ?;
-	`
-	statement, err := db.Prepare(query)
+		WHERE id = ? AND user_id = ?;
+	`)
 	if err != nil {
 		return
 	}
-	_, err = statement.Exec(transaction.Amount, transaction.Category.Id, transaction.Name, transaction.Summary,
-		transaction.TransactionDate, transaction.Id)
-	return
-}
-
-func changeDateFormat(destFormat string, sourceFormat string, sourceDate string) (destDate string) {
-	t, err := time.Parse(sourceFormat, sourceDate)
-	if err != nil {
-		return
-	}
-	destDate = t.Format(destFormat)
+	_, err = statement.Exec(t.Amount, t.Category.Id, t.Name, t.Summary, t.TransactionDate, t.Id, t.UserId)
 	return
 }
