+++ app/transaction_template/transaction_template.go
@@ -1,9 +1,10 @@
 package transaction_template
 
 import (
-	"bitbucket.org/zanvd/accountant/category"
 	"database/sql"
 	"log"
+
+	"bitbucket.org/zanvd/accountant/category"
 )
 
 type TransactionTemplate struct {
@@ -11,6 +12,7 @@ type TransactionTemplate struct {
 	Category category.Category
 	Name     string
 	Position int
+	UserId   int
 }
 
 func CreateTransactionTemplateTable(db *sql.DB) {
@@ -20,11 +22,16 @@ func CreateTransactionTemplateTable(db *sql.DB) {
 			category_id INT DEFAULT NULL,
 			name VARCHAR(30) NOT NULL,
 			position INT NOT NULL DEFAULT 1,
+			user_id INT NOT NULL,
 			PRIMARY KEY (id),
-			FOREIGN KEY category_id_idx (category_id)
-			REFERENCES categories (id)
-			ON DELETE RESTRICT
-			ON UPDATE NO ACTION
+			FOREIGN KEY fk_category_id (category_id)
+				REFERENCES categories (id)
+				ON DELETE RESTRICT
+				ON UPDATE NO ACTION,
+			FOREIGN KEY fk_user_id (user_id)
+				REFERENCES users (id)
+				ON DELETE CASCADE
+				ON UPDATE NO ACTION
 		);
 	`
 	log.Println("Creating the transaction templates table.")
@@ -38,74 +45,76 @@ func CreateTransactionTemplateTable(db *sql.DB) {
 	log.Println("Transaction templates table created.")
 }
 
-func DeleteTransactionTemplate(db *sql.DB, id int) (err error) {
-	_, err = db.Exec("DELETE FROM transaction_templates WHERE id = ?", id)
+func DeleteTransactionTemplate(db *sql.DB, id int, uid int) (err error) {
+	_, err = db.Exec("DELETE FROM transaction_templates WHERE id = ? AND user_id = ?;", id, uid)
 	return
 }
 
-func GetTransactionTemplate(db *sql.DB, id int) (transactionTemplate TransactionTemplate, err error) {
+func GetTransactionTemplate(db *sql.DB, id int, uid int) (tt TransactionTemplate, err error) {
 	query := `
-		SELECT tt.id, tt.name, tt.position, c.id, c.color, c.name, c.text_color
+		SELECT tt.id, tt.name, tt.position, tt.user_id, c.id, c.color, c.name, c.text_color
 		FROM transaction_templates tt
 		LEFT JOIN categories c ON c.id = tt.category_id
-		WHERE tt.id = ?;
+		WHERE tt.id = ? AND tt.user_id = ?;
 	`
-	row := db.QueryRow(query, id)
-	err = row.Scan(&transactionTemplate.Id, &transactionTemplate.Name, &transactionTemplate.Position,
-		&transactionTemplate.Category.Id, &transactionTemplate.Category.Color, &transactionTemplate.Category.Name,
-		&transactionTemplate.Category.TextColor)
+	row := db.QueryRow(query, id, uid)
+	err = row.Scan(
+		&tt.Id, &tt.Name, &tt.Position, &tt.UserId,
+		&tt.Category.Id, &tt.Category.Color, &tt.Category.Name, &tt.Category.TextColor,
+	)
 	return
 }
 
-func GetTransactionTemplates(db *sql.DB, orderByPosition bool) (transactionTemplates []TransactionTemplate, err error) {
+func GetTransactionTemplates(db *sql.DB, orderByPosition bool, uid int) (tts []TransactionTemplate, err error) {
 	query := `
-		SELECT tt.id, tt.name, c.id, c.color, c.name, c.text_color
+		SELECT tt.id, tt.name, tt.position, tt.user_id, c.id, c.color, c.name, c.text_color
 		FROM transaction_templates tt
 		LEFT JOIN categories c ON c.id = tt.category_id
+		WHERE tt.user_id = ?
 	`
 	if orderByPosition {
-		query += " ORDER BY position ASC;"
+		query += " ORDER BY tt.position ASC;"
 	} else {
 		query += ";"
 	}
-	rows, err := db.Query(query)
+	rows, err := db.Query(query, uid)
 	if err != nil {
 		return
 	}
 	defer rows.Close()
 	for rows.Next() {
-		transactionTemplate := TransactionTemplate{}
-		if err = rows.Scan(&transactionTemplate.Id, &transactionTemplate.Name, &transactionTemplate.Category.Id,
-			&transactionTemplate.Category.Color, &transactionTemplate.Category.Name,
-			&transactionTemplate.Category.TextColor); err != nil {
+		tt := TransactionTemplate{}
+		if err = rows.Scan(
+			&tt.Id, &tt.Name, &tt.Position, &tt.UserId,
+			&tt.Category.Id, &tt.Category.Color, &tt.Category.Name, &tt.Category.TextColor,
+		); err != nil {
 			return
 		}
-		transactionTemplates = append(transactionTemplates, transactionTemplate)
+		tts = append(tts, tt)
 	}
 	return
 }
 
-func InsertTransactionTemplate(db *sql.DB, transactionTemplate TransactionTemplate) (err error) {
-	query := "INSERT INTO transaction_templates(category_id, name, position) VALUES (?, ?, ?)"
-	statement, err := db.Prepare(query)
+func InsertTransactionTemplate(db *sql.DB, tt TransactionTemplate) (err error) {
+	statement, err := db.Prepare(
+		"INSERT INTO transaction_templates(category_id, name, position, user_id) VALUES (?, ?, ?, ?);",
+	)
 	if err != nil {
 		return
 	}
-	_, err = statement.Exec(transactionTemplate.Category.Id, transactionTemplate.Name, transactionTemplate.Position)
+	_, err = statement.Exec(tt.Category.Id, tt.Name, tt.Position, tt.UserId)
 	return
 }
 
-func UpdateTransactionTemplate(db *sql.DB, transactionTemplate TransactionTemplate) (err error) {
-	query := `
+func UpdateTransactionTemplate(db *sql.DB, tt TransactionTemplate) (err error) {
+	statement, err := db.Prepare(`
 		UPDATE transaction_templates
 		SET category_id = ?, name = ?, position = ?
-		WHERE id = ?;
-	`
-	statement, err := db.Prepare(query)
+		WHERE id = ? AND user_id = ?;
+	`)
 	if err != nil {
 		return
 	}
-	_, err = statement.Exec(
-		transactionTemplate.Category.Id, transactionTemplate.Name, transactionTemplate.Position, transactionTemplate.Id)
+	_, err = statement.Exec(tt.Category.Id, tt.Name, tt.Position, tt.Id, tt.UserId)
 	return
 }
